+++ packages/rpc_dart_blob/lib/src/adapters/_index.dart
@@ -1,2 +1,3 @@
 export 'i_blob_storage_adapter.dart';
 export 'sqlite_blob_storage_adapter.dart';
+export 's3_blob_storage_adapter.dart';